Clean up effort (#50)
* Clean up effort

This is a clean up effort that gets rid of references to relative directories. Relative directories
are considered bad practice and break easily. Thusly they have been replaced by including paths in
the qmake build files.

* Add PROJECT_ROOT and SOURCE_ROOT to project files
* kicad_global.h was eliminated by defining KICAD_EXPORT in project files
* [[fallthrough]]; was added where it was needed
* Deprecated "endl" replaced with "Qt::endl"
* Use static function when checking if a file exists
* Use pkg-config ability of qmake (if package exists)

* Fixed Qt Signal connection for selectAll

The code attempted to connect the selectAll slot using bool as an argument
but it doesn't have one. The signal argument is optional so it was removed.